Move property tests into test package
